Update patch set 1
Patch Set 1:

More importantly, we give +2 only when we've reviewed the change, and I didn't get a chance to look at this again until today, which I did.

Left some comments on http://review.cyanogenmod.org/98370 which are applicable to shinano too, the rest of the changes (to the individual device repos) look good though.

Patch-set: 1